+++ trunk/ks-multiply.c
@@ -146,7 +146,7 @@ TMulKS (listz_t b, unsigned int n, listz_t a, unsigned int m,
   mp_size_t an, bn, cn;
   int ret = 0; /* default return value */
 #ifdef DEBUG
-  unsigned int st = cputime ();
+  long st = cputime ();
   fprintf (ECM_STDOUT, "n=%u m=%u l=%u bits=%u n*bits=%u: ", n, m, l,
 	   mpz_sizeinbase (modulus, 2), n * mpz_sizeinbase (modulus, 2));
 #endif
